requirements.txt
Django==1.10.3
django-tinymce==2.4.0
djangorestframework==3.5.3
psycopg2==2.6.2
